LU-4684 migrate: migrate striped directory
Migrate striped directory in below steps:
1. create target object if needed: if source is directory, a
   target object is always created, otherwise if source is
   already located on the target MDT, or source still has
   link on source MDT, then skip creating.
	a) if source is directory, detach source stripes and
	   attach them to target.
	b) migrate source xattrs to target.
	c) if source is regular file, update PFID to target
	   fid.
	d) update fid to target for all links of source
2. update namespace
	a) migrate dirent from source parent to target parent.
	b) update linkea parent fid to target parent.
	c) destroy source object.

This implementation improves following fields:
1. all involved objects are locked to avoid race.
2. directory migration doesn't migrate its dir entries, instead
   it's done in each sub file migration, this avoids timeout in
   migrating dir entries for large directory, and also avoids
   touching dir entries without lock.
3. file/dir is migrated in one transaction, so migrate recovery
   is the same as others.
4. migrating directory can be accessed (modifiable) like normal
   directory.
5. if migration of sub files under a directory fails, user can
   redo migrate to finish migration of this directory.

Disable migrate in racer.sh.

Reviewed-on: https://review.whamcloud.com/31427
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/tests/recovery-small.sh
@@ -2109,27 +2109,27 @@ test_110f () {
 run_test 110f "remove remote directory: drop slave rep"
 
 test_110g () {
-	[[ $(lustre_version_code $SINGLEMDS) -ge $(version_code 2.6.57) ]] ||
-		{ skip "Need MDS version at least 2.6.57"; return 0; }
+	[[ $(lustre_version_code $SINGLEMDS) -ge $(version_code 2.11.0) ]] ||
+		{ skip "Need MDS version at least 2.11.0"; return 0; }
 
 	[ $MDSCOUNT -lt 2 ] && skip "needs >= 2 MDTs" && return 0
-	local remote_dir=$DIR/$tdir/remote_dir
-	local mdtidx=1
 
-	mkdir -p $remote_dir
+	mkdir -p $DIR/$tdir
+	touch $DIR/$tdir/$tfile
 
-	createmany -o $remote_dir/f 100
+	# OBD_FAIL_MDS_REINT_NET_REP	0x119
+	do_facet mds1 $LCTL set_param fail_loc=0x119
+	$LFS migrate -m 1 $DIR/$tdir &
+	migrate_pid=$!
+	sleep 5
+	do_facet mds1 $LCTL set_param fail_loc=0
+	wait $migrate_pid
 
-	#define OBD_FAIL_MIGRATE_NET_REP		0x1800
-	do_facet mds$mdtidx lctl set_param fail_loc=0x1800
-	$LFS migrate -m $mdtidx $remote_dir || error "migrate failed"
-	do_facet mds$mdtidx lctl set_param fail_loc=0x0
-
-	for file in $(find $remote_dir); do
-		mdt_index=$($LFS getstripe -m $file)
-		[ $mdt_index == $mdtidx ] ||
-			error "$file is not on MDT${mdtidx}"
-	done
+	local mdt_index
+	mdt_index=$($LFS getstripe -m $DIR/$tdir)
+	[ $mdt_index == 1 ] || error "$tdir is not on MDT1"
+	mdt_index=$($LFS getstripe -m $DIR/$tdir/$tfile)
+	[ $mdt_index == 1 ] || error "$tfile is not on MDT1"
 
 	rm -rf $DIR/$tdir || error "rmdir failed"
 }
